Moved Chess into frontend
