+++ frontend/gamegurureact/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './Home';
-import Chess from './Chess';
+import Chessboard from './Chess/Chessboard';
 import Yahtzee from './Yahtzee';
 import Mastermind from './Mastermind';
 
@@ -10,7 +10,7 @@ function App() {
     <Router>
       <div>
         <Routes>
-          <Route path="/chess" element={<Chess />} />
+          <Route path="/chess" element={<Chessboard />} />
           <Route path="/yahtzee" element={<Yahtzee />} />
           <Route path="/mastermind" element={<Mastermind />} />
           <Route path="/" element={<Home />} />
