--- frontend/gamegurureact/src/Chess.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Chess = () => {
-  return (
-    <div>
-      <h1 style={{ textAlign: 'center' }}>Chess</h1>
-      <p style={{ textAlign: 'center' }}>
-        <a href="/">Back to Home</a>
-      </p>
-    </div>
-  );
-};
-
-export default Chess;
