+++ frontend/gamegurureact/src/assets/bbishop.png
