+++ frontend/gamegurureact/src/assets/bpawn.png
