+++ frontend/gamegurureact/src/assets/bqueen.png
