+++ frontend/gamegurureact/src/assets/wking.png
