+++ frontend/gamegurureact/src/assets/wpawn.png
