+++ frontend/gamegurureact/src/assets/wqueen.png
