+++ frontend/gamegurureact/src/assets/wrook.png
