--- src/Yahtzee.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Yahtzee = () => {
-  return (
-    <div>
-      <h1 style={{ textAlign: 'center' }}>Yahtzee</h1>
-      <p style={{ textAlign: 'center' }}>
-        <a href="/">Back to Home</a>
-      </p>
-    </div>
-  );
-};
-
-export default Yahtzee;
